Add method name() to the *DataSource classes
Summary:
This method provides human-readable name that supposed to be used in
UI elements. One of the usage cases is to give the user an ability to
work with individual data sources (e.g., check data, activate/
deactivate, check description, and so on).
This method is not pure virtual because there are might be some data
sources that are not good for individual processing by the user. For
example, trackers of individual QML properties.

Reviewers: vkrause

Reviewed By: vkrause

Differential Revision: https://phabricator.kde.org/D19756