Add method to check whether a data source is active or not
Summary:
"Active" state means that all data collected by the data source will be
sent to a server.
This change aims to give a user more control of what data they want to
send. Without this functionality the user might decide to reduce
telemetry level because of one small part of it. This means that we can
collect less statistic without a really good reason.

Reviewers: vkrause

Differential Revision: https://phabricator.kde.org/D19757