Add static build support.
Summary:
It's possible to build the framework as a set of static libraries now.
Use -DBUILD_SHARED_LIBS=OFF to enable static build.

The following libraries will be created:
libKUserFeedbackCore.a
libKUserFeedbackWidgets.a
libKUserFeedbackCommon.a

You have to keep them in order listed above, when linking your app
against them.

Reviewers: vkrause

Reviewed By: vkrause

Differential Revision: https://phabricator.kde.org/D20575