Revise title and description texts for the different modes
Summary:
Currently, the text in the user feedback KCM is quite intimidating and scary to me for
various reasons:
- The titles do not accurately describe what will actually be sent
- The description text of "don't send anything" is inappropriately a description of everything
- The same slightly awkward "here's what we don't send" text is releated in every description

Overall it does not give me confidence and make me want to submit information.

This patch (hopefully) improves the text by making the titles more descriptive, making the
description text simpler and more user-friendly, ensuring that every bit of description
text is actually describing its setting, and removing the "here's what we don't send" text,
which should be shown only once by the UI that presents this information (e.g. the Feedback
KCM).

Test Plan:
How it looks now in the Feedback KCM (with D26552 applied): {F7875839}

How it looks in UserFeedbackConsole: {F7875840}

Reviewers: #vdg, #plasma, apol, vkrause

Differential Revision: https://phabricator.kde.org/D26551