FeedbackConfigWidget: replace audit log label with button
We never use link labels for interactive parts of UI that aren't a web browser.
A button is much more in line with the rest of KDE's software, as well as having
proper semantics (important for screen readers) and being able to show a disabled
state (important for all users).