tests: allow to disable testOpenGLInfoSource using envvar
testOpenGLInfoSource actually tries to instantiate an OpenGL context to
get its details; this may not be always possible in a reliable way,
due to hardware, architecture, software (e.g. Mesa), and so on.

Hence, to still enable the execution of the tests, add the possibility
to skip that specific test when a certain environment variable is set.

(cherry picked from commit 9bb9187846a19c0f366ab9ce6369edd25086673d)