Properly handling large timeouts for encouragement timer
If the last encouragement was recently then 30 days later is a long
time from now. Particularly enough so that when converting it to int
it will overflow and result in a negative number.

We then pass that negative value to QTimer. Since Qt 6.10 that results
in a timeout of 1ms, which causes the encouragement to trigger immediately.

To avoid that make sure to use 64bit integers for the handling, and use
QChronoTimer, which can properly handle timeouts too large for 32bit int

BUG: 511102